+++ test/smokes/pmd_cpd/expectations.rb
@@ -24,35 +24,7 @@ s.add_test(
             id: "069e8856ac66563b0c6e50b615209face9958fa7", path: "app.java",
             start_line: 31, start_column: 1, end_line: 58, end_column: 1
           }
-        ],
-        codefragment: %[public class Main {
-  public static void main(String[] args) {
-    int day = 4;
-    switch (day) {
-      case 1:
-        System.out.println("Monday");
-        break;
-      case 2:
-        System.out.println("Tuesday");
-        break;
-      case 3:
-        System.out.println("Wednesday");
-        break;
-      case 4:
-        System.out.println("Thursday");
-        break;
-      case 5:
-        System.out.println("Friday");
-        break;
-      case 6:
-        System.out.println("Saturday");
-        break;
-      case 7:
-        System.out.println("Sunday");
-        break;
-    }
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "b4aa7ff94732722c55d79ff3788d3d5f22b197b2", original_line: 31, final_line: 31
@@ -76,35 +48,7 @@ s.add_test(
             id: "069e8856ac66563b0c6e50b615209face9958fa7", path: "app.java",
             start_line: 31, start_column: 1, end_line: 58, end_column: 1
           }
-        ],
-        codefragment: %[public class Main {
-  public static void main(String[] args) {
-    int day = 4;
-    switch (day) {
-      case 1:
-        System.out.println("Monday");
-        break;
-      case 2:
-        System.out.println("Tuesday");
-        break;
-      case 3:
-        System.out.println("Wednesday");
-        break;
-      case 4:
-        System.out.println("Thursday");
-        break;
-      case 5:
-        System.out.println("Friday");
-        break;
-      case 6:
-        System.out.println("Saturday");
-        break;
-      case 7:
-        System.out.println("Sunday");
-        break;
-    }
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "b4aa7ff94732722c55d79ff3788d3d5f22b197b2", original_line: 1, final_line: 1
@@ -157,35 +101,7 @@ s.add_test(
             id: "f721b99980175debf0b6e860af6e29b43b4509b3", path: "lib/foo/bar.java",
             start_line: 4, start_column: 1, end_line: 31, end_column: 1
           }
-        ],
-        codefragment: %[public class Main {
-  public static void main(String[] args) {
-    int day = 4;
-    switch (day) {
-      case 1:
-        System.out.println("Monday");
-        break;
-      case 2:
-        System.out.println("Tuesday");
-        break;
-      case 3:
-        System.out.println("Wednesday");
-        break;
-      case 4:
-        System.out.println("Thursday");
-        break;
-      case 5:
-        System.out.println("Friday");
-        break;
-      case 6:
-        System.out.println("Saturday");
-        break;
-      case 7:
-        System.out.println("Sunday");
-        break;
-    }
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "b4aa7ff94732722c55d79ff3788d3d5f22b197b2", original_line: 4, final_line: 4
@@ -209,35 +125,7 @@ s.add_test(
             id: "f721b99980175debf0b6e860af6e29b43b4509b3", path: "lib/foo/bar.java",
             start_line: 4, start_column: 1, end_line: 31, end_column: 1
           }
-        ],
-        codefragment: %[public class Main {
-  public static void main(String[] args) {
-    int day = 4;
-    switch (day) {
-      case 1:
-        System.out.println("Monday");
-        break;
-      case 2:
-        System.out.println("Tuesday");
-        break;
-      case 3:
-        System.out.println("Wednesday");
-        break;
-      case 4:
-        System.out.println("Thursday");
-        break;
-      case 5:
-        System.out.println("Friday");
-        break;
-      case 6:
-        System.out.println("Saturday");
-        break;
-      case 7:
-        System.out.println("Sunday");
-        break;
-    }
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "b4aa7ff94732722c55d79ff3788d3d5f22b197b2", original_line: 1, final_line: 1
@@ -269,21 +157,7 @@ s.add_test(
             id: "f187a6d7cc29608a9e1e931d5b29e8d489d7fdbd", path: "src/app.rb",
             start_line: 3, start_column: 1, end_line: 16, end_column: 3
           }
-        ],
-        codefragment: %[def show_status_tank capacity
-  case capacity
-  when 0
-    "You ran out of gas."
-  when 1..20
-    "The tank is almost empty. Quickly, find a gas station!"
-  when 21..70
-    "You should be ok for now."
-  when 71..100
-    "The tank is almost full."
-  else
-    "Error: capacity has an invalid value (\#{capacity})"
-  end
-end]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "cb9bf4628574bad099e82e3b56c908af7ae30231", original_line: 5, final_line: 5
@@ -307,21 +181,7 @@ end]
             id: "f187a6d7cc29608a9e1e931d5b29e8d489d7fdbd", path: "src/app.rb",
             start_line: 3, start_column: 1, end_line: 16, end_column: 3
           }
-        ],
-        codefragment: %[def show_status_tank capacity
-  case capacity
-  when 0
-    "You ran out of gas."
-  when 1..20
-    "The tank is almost empty. Quickly, find a gas station!"
-  when 21..70
-    "You should be ok for now."
-  when 71..100
-    "The tank is almost full."
-  else
-    "Error: capacity has an invalid value (\#{capacity})"
-  end
-end]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "cb9bf4628574bad099e82e3b56c908af7ae30231", original_line: 3, final_line: 3
@@ -353,10 +213,7 @@ s.add_test(
             id: "3b57b2623d8a19a1a050be0ffe54d21d22cac853", path: "src/fizzbuzz_2.c",
             start_line: 11, start_column: 16, end_line: 13, end_column: 13
           }
-        ],
-        codefragment: %[    } else if (i % 3 == 0) {
-      printf("Fizz\\n");
-    } else if (i % 5 == 0) {]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "c2189bfdb7e8b36e6d4cadc98bc1f0c9adf0c65b", original_line: 11, final_line: 11
@@ -380,11 +237,7 @@ s.add_test(
             id: "e33562290c84ad67d2a2937738674ca7b8fb0397", path: "src/fizzbuzz_2.c",
             start_line: 13, start_column: 16, end_line: 16, end_column: 18
           }
-        ],
-        codefragment: %[    } else if (i % 5 == 0) {
-      printf("Buzz\\n");
-    } else {
-      printf("%d\\n", i);]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "c6dc0e3ff36eaf3fffc36618262b81d07e266dc3", original_line: 12, final_line: 12
@@ -408,10 +261,7 @@ s.add_test(
             id: "cd80f7d3fa55dde6def3d4114931f102038a6abc", path: "src/fizzbuzz_2.c",
             start_line: 9, start_column: 23, end_line: 11, end_column: 13
           }
-        ],
-        codefragment: %[    if (i % 3 == 0 && i % 5 == 0) {
-      printf("Fizz, Buzz\\n");
-    } else if (i % 3 == 0) {]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "59601bb12fa954e3157d45d1a562240d634a572a", original_line: 8, final_line: 8
@@ -435,10 +285,7 @@ s.add_test(
             id: "cd80f7d3fa55dde6def3d4114931f102038a6abc", path: "src/fizzbuzz_2.c",
             start_line: 9, start_column: 23, end_line: 11, end_column: 13
           }
-        ],
-        codefragment: %[    if (i % 3 == 0 && i % 5 == 0) {
-      printf("Fizz, Buzz\\n");
-    } else if (i % 3 == 0) {]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "cb3593813b1bc80eccbb9c6921dc3d70944a844b", original_line: 9, final_line: 9
@@ -462,11 +309,7 @@ s.add_test(
             id: "e33562290c84ad67d2a2937738674ca7b8fb0397", path: "src/fizzbuzz_2.c",
             start_line: 13, start_column: 16, end_line: 16, end_column: 18
           }
-        ],
-        codefragment: %[    } else if (i % 5 == 0) {
-      printf("Buzz\\n");
-    } else {
-      printf("%d\\n", i);]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "4001228b9db17d5b70e2867a83fc5e970aa14338", original_line: 13, final_line: 13
@@ -490,10 +333,7 @@ s.add_test(
             id: "3b57b2623d8a19a1a050be0ffe54d21d22cac853", path: "src/fizzbuzz_2.c",
             start_line: 11, start_column: 16, end_line: 13, end_column: 13
           }
-        ],
-        codefragment: %[    } else if (i % 3 == 0) {
-      printf("Fizz\\n");
-    } else if (i % 5 == 0) {]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "2fa7627f128681125e2074d6f40d0011e1dd3641", original_line: 10, final_line: 10
@@ -525,18 +365,7 @@ s.add_test(
             id: "7e98a5dcfadd5f7617d390c6ab4dbf2b3008fda8", path: "src/fizzbuzz.c",
             start_line: 20, start_column: 3, end_line: 30, end_column: 3
           }
-        ],
-        codefragment: %[  for (i = 1; i <= 100; i++) {
-    if (i % 3 == 0 && i % 5 == 0) {
-      printf("Fizz, Buzz\\n");
-    } else if (i % 3 == 0) {
-      printf("Fizz\\n");
-    } else if (i % 5 == 0) {
-      printf("Buzz\\n");
-    } else {
-      printf("%d\\n", i);
-    }
-  }]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "3f0d90fd75bd48933ec659984d4fc661105e9d82", original_line: 7, final_line: 7
@@ -560,18 +389,7 @@ s.add_test(
             id: "7e98a5dcfadd5f7617d390c6ab4dbf2b3008fda8", path: "src/fizzbuzz.c",
             start_line: 20, start_column: 3, end_line: 30, end_column: 3
           }
-        ],
-        codefragment: %[  for (i = 1; i <= 100; i++) {
-    if (i % 3 == 0 && i % 5 == 0) {
-      printf("Fizz, Buzz\\n");
-    } else if (i % 3 == 0) {
-      printf("Fizz\\n");
-    } else if (i % 5 == 0) {
-      printf("Buzz\\n");
-    } else {
-      printf("%d\\n", i);
-    }
-  }]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "3f0d90fd75bd48933ec659984d4fc661105e9d82", original_line: 20, final_line: 20
@@ -603,31 +421,7 @@ s.add_test(
             id: "37f517be4959df9fb77a38308b522a24d674c0aa", path: "foo/こんにちは.sjis.cs",
             start_line: 1, start_column: 1, end_line: 24, end_column: 1
           }
-        ],
-        codefragment: %[using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "da8982440e79f7642984edea9224861708e6da31", original_line: 1, final_line: 1
@@ -651,31 +445,7 @@ namespace MyApp
             id: "37f517be4959df9fb77a38308b522a24d674c0aa", path: "foo/こんにちは.sjis.cs",
             start_line: 1, start_column: 1, end_line: 24, end_column: 1
           }
-        ],
-        codefragment: %[using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "da8982440e79f7642984edea9224861708e6da31", original_line: 3, final_line: 3
@@ -722,22 +492,7 @@ s.add_test(
             id: "408776ef9c22448f623aae1a15f82cab1d133846", path: "bar/こんにちは.1.eucjp.cs",
             start_line: 14, start_column: 1, end_line: 28, end_column: 1
           }
-        ],
-        codefragment: %[namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ccf27261b7163c516af19429667e004914add570", original_line: 16, final_line: 16
@@ -761,22 +516,7 @@ s.add_test(
             id: "408776ef9c22448f623aae1a15f82cab1d133846", path: "bar/こんにちは.1.eucjp.cs",
             start_line: 14, start_column: 1, end_line: 28, end_column: 1
           }
-        ],
-        codefragment: %[namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ccf27261b7163c516af19429667e004914add570", original_line: 14, final_line: 14
@@ -818,47 +558,7 @@ s.add_test(
             id: "e5d36afbb6911b96a9e52d77388c9cad9c3cedf0", path: "MyApp.java",
             start_line: 1, start_column: 1, end_line: 40, end_column: 1
           }
-        ],
-        codefragment: %[public class MyAppJp {
-  public static void main(String[] args) {
-    String str;
-    int day = 5;
-    switch (day) {
-      case 1:
-        str = "月曜日";
-        break;
-      case 2:
-        str = "火曜日";
-        break;
-      case 3:
-        str = "水曜日";
-        break;
-      case 4:
-        str = "木曜日";
-        break;
-      case 5:
-        str = "金曜日";
-        break;
-      case 6:
-        str = "土曜日";
-        break;
-      case 7:
-        str = "日曜日";
-        break;
-      default:
-        str = null;
-        break;
-    }
-
-    System.out.println(str);
-  }
-
-  @Deprecated
-  private static void Foo(String text)
-  {
-    System.out.println(text);
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "bc588a44b0d2d31c9d7ead684d8edfceffd513aa", original_line: 3, final_line: 3
@@ -882,47 +582,7 @@ s.add_test(
             id: "e5d36afbb6911b96a9e52d77388c9cad9c3cedf0", path: "MyApp.java",
             start_line: 1, start_column: 1, end_line: 40, end_column: 1
           }
-        ],
-        codefragment: %[public class MyAppJp {
-  public static void main(String[] args) {
-    String str;
-    int day = 5;
-    switch (day) {
-      case 1:
-        str = "月曜日";
-        break;
-      case 2:
-        str = "火曜日";
-        break;
-      case 3:
-        str = "水曜日";
-        break;
-      case 4:
-        str = "木曜日";
-        break;
-      case 5:
-        str = "金曜日";
-        break;
-      case 6:
-        str = "土曜日";
-        break;
-      case 7:
-        str = "日曜日";
-        break;
-      default:
-        str = null;
-        break;
-    }
-
-    System.out.println(str);
-  }
-
-  @Deprecated
-  private static void Foo(String text)
-  {
-    System.out.println(text);
-  }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "0ba457a9706d0017c98765a0e1ad6b43c4635564", original_line: 1, final_line: 1
@@ -954,22 +614,7 @@ s.add_test(
             id: "0d0169de6995b44aca9b1860372a6428bf67faa7", path: "src/bar/こんにちは.1.sjis.cs",
             start_line: 12, start_column: 1, end_line: 26, end_column: 1
           }
-        ],
-        codefragment: %[namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ccf27261b7163c516af19429667e004914add570", original_line: 12, final_line: 12
@@ -993,22 +638,7 @@ s.add_test(
             id: "0d0169de6995b44aca9b1860372a6428bf67faa7", path: "src/bar/こんにちは.1.sjis.cs",
             start_line: 12, start_column: 1, end_line: 26, end_column: 1
           }
-        ],
-        codefragment: %[namespace MyApp
-{
-    public class Hello
-    {
-        public static void Main()
-        {
-            今日は世界();
-        }
-
-        private static void 今日は世界()
-        {
-            Console.WriteLine("今日は, 世界!");
-        }
-    }
-}]
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ccf27261b7163c516af19429667e004914add570", original_line: 12, final_line: 12
@@ -1054,8 +684,7 @@ s.add_test(
             id: "f596d2f766fb7b912331c3691b493b5d028f8ce3", path: "foo/hello_cpp.cpp",
             start_line: 3, start_column: 1, end_line: 17, end_column: 1
           }
-        ],
-        codefragment: /void print_hello\(\);/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "c5a6d5c927086556e95ecdbf3a2bc806d3e51162", original_line: 3, final_line: 3
@@ -1079,8 +708,7 @@ s.add_test(
             id: "2313e91e4b39484760de0c6a7de02d9fcd7db381", path: "foo/hello_ecmascript.js",
             start_line: 1, start_column: 1, end_line: 5, end_column: 14
           }
-        ],
-        codefragment: /function print_hello\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "5151e3c05aa07d1ec78fff69e72c058b82b4b697", original_line: 1, final_line: 1
@@ -1104,8 +732,7 @@ s.add_test(
             id: "59d1a6ee4579c0bb7c7abecf5c0aaa006988b5e4", path: "foo/hello_java.java",
             start_line: 1, start_column: 1, end_line: 13, end_column: 1
           }
-        ],
-        codefragment: /class TestJp/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "8c2df874c382428b28f2aabe6d2a0eba05f51aed", original_line: 1, final_line: 1
@@ -1129,8 +756,7 @@ s.add_test(
             id: "3f18338e7a3124257c7b1424604e3386e6808bc4", path: "foo/hello_php.php",
             start_line: 1, start_column: nil, end_line: 8, end_column: nil
           }
-        ],
-        codefragment: /<\?php/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "9166a7b7093b6ef318e436c6b16866e360ab4381", original_line: 1, final_line: 1
@@ -1154,8 +780,7 @@ s.add_test(
             id: "551c2990310dfc0c56a00a83d7e2b3010de61732", path: "foo/hello_kotlin.kt",
             start_line: 1, start_column: 1, end_line: 7, end_column: 1
           }
-        ],
-        codefragment: /fun main\(args: Array<String>\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "eb3f822ef926724389b80800b171ce44e4758cd5", original_line: 1, final_line: 1
@@ -1179,8 +804,7 @@ s.add_test(
             id: "2313e91e4b39484760de0c6a7de02d9fcd7db381", path: "foo/hello_ecmascript.js",
             start_line: 1, start_column: 1, end_line: 5, end_column: 14
           }
-        ],
-        codefragment: /function print_hello\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "5151e3c05aa07d1ec78fff69e72c058b82b4b697", original_line: 1, final_line: 1
@@ -1204,8 +828,7 @@ s.add_test(
             id: "59d1a6ee4579c0bb7c7abecf5c0aaa006988b5e4", path: "foo/hello_java.java",
             start_line: 1, start_column: 1, end_line: 13, end_column: 1
           }
-        ],
-        codefragment: /class TestJp/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "96c0f9c6fc598d7b85fd63000844909ffd2c2cde", original_line: 1, final_line: 1
@@ -1229,8 +852,7 @@ s.add_test(
             id: "7e380f221fcff72d4568a72141e1a29d1acf9ba4", path: "foo/hello_go.go",
             start_line: 1, start_column: 1, end_line: 10, end_column: 1
           }
-        ],
-        codefragment: /package main/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "04eb6f1bdaf51ae48ed5cf0153fad8593467b778", original_line: 1, final_line: 1
@@ -1254,8 +876,7 @@ s.add_test(
             id: "80e9eb7928ca5770ffe588c59fa744f1b197fc68", path: "foo/hello_ruby.rb",
             start_line: 1, start_column: 1, end_line: 24, end_column: 6
           }
-        ],
-        codefragment: /def main\(\)/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "763d80e283cd30bc48a6425fb1adbb97ba090bf5", original_line: 1, final_line: 1
@@ -1279,8 +900,7 @@ s.add_test(
             id: "3f18338e7a3124257c7b1424604e3386e6808bc4", path: "foo/hello_php.php",
             start_line: 1, start_column: nil, end_line: 8, end_column: nil
           }
-        ],
-        codefragment: /<\?php/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "9166a7b7093b6ef318e436c6b16866e360ab4381", original_line: 1, final_line: 1
@@ -1304,8 +924,7 @@ s.add_test(
             id: "80e9eb7928ca5770ffe588c59fa744f1b197fc68", path: "foo/hello_ruby.rb",
             start_line: 1, start_column: 1, end_line: 24, end_column: 6
           }
-        ],
-        codefragment: /def main\(\)/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "763d80e283cd30bc48a6425fb1adbb97ba090bf5", original_line: 1, final_line: 1
@@ -1329,8 +948,7 @@ s.add_test(
             id: "ff532463942827ebfcbe38b1f1c34924421d13ef", path: "foo/hello_cs.cs",
             start_line: 5, start_column: 1, end_line: 16, end_column: 1
           }
-        ],
-        codefragment: /public class Test/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "a6fbaa5beef91bb1bc2b212bf5f9f079263c153a", original_line: 5, final_line: 5
@@ -1354,8 +972,7 @@ s.add_test(
             id: "c78059bc1c128c7f394e24848ba626ea6983bcb8", path: "foo/hello_python.py",
             start_line: 1, start_column: 1, end_line: 7, end_column: 6
           }
-        ],
-        codefragment: /def main\(\):/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "65b414a74c147117d8053894329f9232370aa58f", original_line: 1, final_line: 1
@@ -1379,8 +996,7 @@ s.add_test(
             id: "7e380f221fcff72d4568a72141e1a29d1acf9ba4", path: "foo/hello_go.go",
             start_line: 1, start_column: 1, end_line: 10, end_column: 1
           }
-        ],
-        codefragment: /package main/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "04eb6f1bdaf51ae48ed5cf0153fad8593467b778", original_line: 1, final_line: 1
@@ -1404,8 +1020,7 @@ s.add_test(
             id: "551c2990310dfc0c56a00a83d7e2b3010de61732", path: "foo/hello_kotlin.kt",
             start_line: 1, start_column: 1, end_line: 7, end_column: 1
           }
-        ],
-        codefragment: /fun main\(args: Array<String>\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "eb3f822ef926724389b80800b171ce44e4758cd5", original_line: 1, final_line: 1
@@ -1429,8 +1044,7 @@ s.add_test(
             id: "c78059bc1c128c7f394e24848ba626ea6983bcb8", path: "foo/hello_python.py",
             start_line: 1, start_column: 1, end_line: 7, end_column: 6
           }
-        ],
-        codefragment: /def main\(\):/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "65b414a74c147117d8053894329f9232370aa58f", original_line: 1, final_line: 1
@@ -1454,8 +1068,7 @@ s.add_test(
             id: "c89018422ac15c0d1fc595195903c7c65371fac3", path: "foo/hello_swift.swift",
             start_line: 1, start_column: 1, end_line: 9, end_column: 6
           }
-        ],
-        codefragment: /func main\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "5656b6543b24b9b164b717943f7e56d1cf5954c0", original_line: 1, final_line: 1
@@ -1479,8 +1092,7 @@ s.add_test(
             id: "f596d2f766fb7b912331c3691b493b5d028f8ce3", path: "foo/hello_cpp.cpp",
             start_line: 3, start_column: 1, end_line: 17, end_column: 1
           }
-        ],
-        codefragment: /void print_hello\(\);/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "c5a6d5c927086556e95ecdbf3a2bc806d3e51162", original_line: 3, final_line: 3
@@ -1504,8 +1116,7 @@ s.add_test(
             id: "c89018422ac15c0d1fc595195903c7c65371fac3", path: "foo/hello_swift.swift",
             start_line: 1, start_column: 1, end_line: 9, end_column: 6
           }
-        ],
-        codefragment: /func main\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "5656b6543b24b9b164b717943f7e56d1cf5954c0", original_line: 1, final_line: 1
@@ -1529,8 +1140,7 @@ s.add_test(
             id: "ff532463942827ebfcbe38b1f1c34924421d13ef", path: "foo/hello_cs.cs",
             start_line: 5, start_column: 1, end_line: 16, end_column: 1
           }
-        ],
-        codefragment: /public class Test/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "a6fbaa5beef91bb1bc2b212bf5f9f079263c153a", original_line: 5, final_line: 5
@@ -1562,8 +1172,7 @@ s.add_test(
             id: "8b7407bcff2b251861c4454c59c29184b78750e9", path: "src/foo/hello_dart.dart",
             start_line: 1, start_column: 1, end_line: 7, end_column: 1
           }
-        ],
-        codefragment: /void main\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "a54d689d7e85de2ea6fbc1abfcfab8fb325fb9f6", original_line: 1, final_line: 1
@@ -1587,8 +1196,7 @@ s.add_test(
             id: "266f2b1a5dc72b41758be1aca2ebea206ea4cf24", path: "src/foo/hello_python.py",
             start_line: 1, start_column: 1, end_line: 7, end_column: 6
           }
-        ],
-        codefragment: /def main\(\):/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "65b414a74c147117d8053894329f9232370aa58f", original_line: 1, final_line: 1
@@ -1612,8 +1220,7 @@ s.add_test(
             id: "266f2b1a5dc72b41758be1aca2ebea206ea4cf24", path: "src/foo/hello_python.py",
             start_line: 1, start_column: 1, end_line: 7, end_column: 6
           }
-        ],
-        codefragment: /def main\(\):/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "65b414a74c147117d8053894329f9232370aa58f", original_line: 1, final_line: 1
@@ -1637,8 +1244,7 @@ s.add_test(
             id: "8b7407bcff2b251861c4454c59c29184b78750e9", path: "src/foo/hello_dart.dart",
             start_line: 1, start_column: 1, end_line: 7, end_column: 1
           }
-        ],
-        codefragment: /void main\(\) \{/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "a54d689d7e85de2ea6fbc1abfcfab8fb325fb9f6", original_line: 1, final_line: 1
@@ -1662,8 +1268,7 @@ s.add_test(
             id: "b4b263f2acec2d1721e04791c768966feedb8080", path: "src/foo/hello_lua.lua",
             start_line: 1, start_column: 1, end_line: 9, end_column: 6
           }
-        ],
-        codefragment: /function main\(\)/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ee2db43a54387611d5dc5ca3899010dd4c8f8365", original_line: 1, final_line: 1
@@ -1687,8 +1292,7 @@ s.add_test(
             id: "b4b263f2acec2d1721e04791c768966feedb8080", path: "src/foo/hello_lua.lua",
             start_line: 1, start_column: 1, end_line: 9, end_column: 6
           }
-        ],
-        codefragment: /function main\(\)/
+        ]
       },
       git_blame_info: {
         commit: :_, line_hash: "ee2db43a54387611d5dc5ca3899010dd4c8f8365", original_line: 1, final_line: 1